Remove unnecessary dependency on Requests
This commit also expands the example and plugin documentation, reformats
the docstrings, and employs f-strings to build the URLs (since Fauxmo
requires >=3.6 regardless).

Removing the dependency on Requests (and the previous dependency on
homeassistant) is especially welcome, as now this can reasonably be
included in the main Fauxmo repo (which I prefer to keep dependency
free).

Cheers, @jld3103, thanks for getting this started!